refactor: move icon.tsx out of public
